+++ tamr_client/attributes/type_alias.py
@@ -1,0 +1,23 @@
+import tamr_client as tc
+from tamr_client.attributes.attribute_type import Array, DOUBLE, Record, STRING
+
+DEFAULT = Array(STRING)
+
+GEOSPATIAL = Record(
+    attributes=(
+        tc.SubAttribute(name="point", is_nullable=True, type=Array(DOUBLE)),
+        tc.SubAttribute(name="multiPoint", is_nullable=True, type=Array(Array(DOUBLE))),
+        tc.SubAttribute(name="lineString", is_nullable=True, type=Array(Array(DOUBLE))),
+        tc.SubAttribute(
+            name="multiLineString", is_nullable=True, type=Array(Array(Array(DOUBLE)))
+        ),
+        tc.SubAttribute(
+            name="polygon", is_nullable=True, type=Array(Array(Array(DOUBLE)))
+        ),
+        tc.SubAttribute(
+            name="multiPolygon",
+            is_nullable=True,
+            type=Array(Array(Array(Array(DOUBLE)))),
+        ),
+    )
+)