+++ tamr_client/json_dict.py
@@ -1,0 +1,4 @@
+# taken from https://github.com/python/typing/issues/182
+from typing import Any, Dict
+
+JsonDict = Dict[str, Any]