+++ tamr_unify_client/__init__.py
@@ -1,6 +1,5 @@
+# flake8: noqa
+
 import logging
 
 from tamr_unify_client.client import Client
-
-# https://docs.python-guide.org/writing/logging/#logging-in-a-library
-logging.getLogger(__name__).addHandler(logging.NullHandler())