+++ tamr_unify_client/auth/__init__.py
@@ -1,2 +1,4 @@
+# flake8: noqa
+
 from tamr_unify_client.auth.token import TokenAuth
 from tamr_unify_client.auth.username_password import UsernamePasswordAuth