+++ tasks.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 from invoke import task
 
 
@@ -10,6 +12,22 @@ def lint(c):
 def format(c, fix=False):
     check = "" if fix else "--check"
     c.run(f"poetry run black {check} .", echo=True, pty=True)
+
+
+@task
+def typecheck(c, warn=True):
+    repo = Path(".")
+    tc = repo / "tamr_client"
+    tests = repo / "tests"
+    pkgs = [
+        tc / "attributes",
+        tests / "attributes",
+        tc / "datasets",
+        tests / "datasets",
+    ]
+    for pkg in pkgs:
+        pyfiles = " ".join(str(pyfile) for pyfile in pkg.glob("**/*.py"))
+        c.run(f"poetry run mypy {pyfiles}", echo=True, pty=True, warn=warn)
 
 
 @task